+++ app/src/main/java/com/devenlucaz/doscom/character/CharacterState.kt
@@ -1,0 +1,16 @@
+package com.devenlucaz.doscom.character
+
+enum class CharacterState {
+    IDLE_BLINK,
+    IDLE_BOB,
+    IDLE_LOOK_LEFT,
+    IDLE_LOOK_RIGHT,
+    WALK_RIGHT,
+    WALK_LEFT,
+    POINT,
+    LISTEN,
+    REACT_WAVE,
+    REACT_WORRY,
+    REACT_HAPPY,
+    TALK
+}